requirements.txt
fastapi
uvicorn
requests
spacy
pdfplumber
pydantic
python-dotenv
pytest
transformers
